+++ src/router/index.ts
@@ -1,5 +1,5 @@
 import { createRouter, createWebHistory } from "vue-router";
-import HomeView from "@/views/HomeView.vue";
+import HomePage from "@/views/HomePage.vue";
 const router = createRouter({
   history: createWebHistory(import.meta.env.BASE_URL),
   routes: [
@@ -7,7 +7,7 @@ const router = createRouter({
       path: "/",
       name: "home",
       alias: "/home",
-      component: HomeView,
+      component: HomePage,
     },
     {
       path: "/contact",
